fix(monitor): Monitor-alive logic, reinitialize drivers
It is confirmed that there are re-entry issues, when accessing
properties of the drivers. The reason is not yet known.

I suspect it has something to do with writing properties with
asyncio, so at some point I may try turning every property
wrappers back into methods.

The problem of monitor got stuck is very likely due to the lock-
release process, and somehow the whole coroutine goes into a deadlock.
For now, I will just delete all devices and re-initialize them.